Progress on basic logic
I now have a text field that can report when new text is entered into it
and I can retrieve the string from that field.  Next up is limiting the
amount of text that can be entered by an amount equal to the target
text, and getting a countdown timer going.